adds gravatar
